% Auto-generated: do not edit by hand
\name{csvboxImporterForDash}

\alias{csvboxImporterForDash}

\title{CsvboxImporterForDash component}

\description{
CsvboxImporterForDash is a component for CSVBox.io integration in Dash.
}

\usage{
csvboxImporterForDash(id=NULL, data=NULL, licenseKey=NULL, result=NULL,
userId=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Named list. Data from the import operation.}

\item{licenseKey}{Character. The license key for CSVBox.io.}

\item{result}{Logical. Result of the import operation.}

\item{userId}{Character. The user ID for CSVBox.io.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

